refactor(groups): extract GroupAttachmentBinder from GroupConversationService
Add the Rev 2.1 group-collaboration implementation plan and begin its Wave R
refactoring workstream: GroupConversationService (4,417 lines) is decomposed
into focused collaborator classes before the plan's ~18 feature items land,
so they get clean homes instead of piling onto an already-oversized file.

R1 step 1 (smallest, pure-move extraction first per the plan's ordering):
move the attachment cluster (materializeAttachments, rehydrateAttachmentsFromStore,
grantAndInjectAttachments) into a new GroupAttachmentBinder. Plain class, not a
CDI bean, constructed per call site — the existing test suite constructs
GroupConversationService directly and attachmentStore is field-injected
specifically to keep that compiling, so the extraction must not change the
constructor signature.

Baselined the full 12-class/470-test GroupConversationService*Test suite
before touching code (all green, 82% instruction / 72% branch JaCoCo) as the
refactor's regression budget. Moved the 12 dedicated attachment tests to a
new focused GroupAttachmentBinderTest testing the extracted class directly.
Re-verified: still 470 tests total, all green, clean compile, clean
formatter/Checkstyle.
