+++ src/test/java/ai/labs/eddi/engine/internal/GroupConversationServiceTest.java
@@ -26,7 +26,6 @@ import ai.labs.eddi.engine.api.IConversationService;
 import ai.labs.eddi.engine.api.IGroupConversationService.GroupDepthExceededException;
 import ai.labs.eddi.engine.api.IGroupConversationService.GroupDiscussionException;
 import ai.labs.eddi.engine.api.IGroupConversationService.GroupDiscussionEventListener;
-import ai.labs.eddi.engine.attachments.IAttachmentStore;
 import ai.labs.eddi.engine.lifecycle.model.ControlSignal;
 import ai.labs.eddi.engine.model.Context;
 import ai.labs.eddi.engine.memory.model.Attachment;
@@ -99,188 +98,10 @@ class GroupConversationServiceTest {
     }
 
     // =================================================================
-    // attachment materialize / grant / inject
+    // attachment materialize / grant / inject — moved to
+    // ai.labs.eddi.engine.internal.groups.GroupAttachmentBinderTest
+    // (Wave R, R1 step 1: logic extracted to GroupAttachmentBinder)
     // =================================================================
-
-    @Nested
-    class Attachments {
-
-        private GroupConversation gc(String id) {
-            var gc = new GroupConversation();
-            gc.setId(id);
-            return gc;
-        }
-
-        @Test
-        void materialize_base64_storesAndBinds() throws Exception {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            when(store.store(any(), eq("image/png"), eq("a.png"), eq("gc-1"), eq(DEFAULT_TENANT)))
-                    .thenReturn(new IAttachmentStore.Attachment("ref-1", "a.png", "image/png", 3, "gc-1"));
-
-            var inline = new Attachment();
-            inline.setMimeType("image/png");
-            inline.setFileName("a.png");
-            inline.setBase64Data(java.util.Base64.getEncoder().encodeToString("png".getBytes()));
-            var gc = gc("gc-1");
-
-            service.materializeAttachments(gc, List.of(inline));
-
-            assertEquals(1, gc.getAttachments().size());
-            assertEquals("ref-1", gc.getAttachments().get(0).getStorageRef());
-        }
-
-        @Test
-        void materialize_url_passesThrough() {
-            service.attachmentStore = mock(IAttachmentStore.class);
-            var url = new Attachment();
-            url.setMimeType("image/png");
-            url.setUrl("https://example.com/y.png");
-            var gc = gc("gc-1");
-
-            service.materializeAttachments(gc, List.of(url));
-
-            assertEquals("https://example.com/y.png", gc.getAttachments().get(0).getUrl());
-        }
-
-        @Test
-        void materialize_noStore_dropsInlineButKeepsUrl() {
-            service.attachmentStore = null;
-            var inline = new Attachment();
-            inline.setBase64Data("x");
-            var url = new Attachment();
-            url.setMimeType("image/png");
-            url.setUrl("https://example.com/y.png");
-            var gc = gc("gc-1");
-
-            service.materializeAttachments(gc, List.of(inline, url));
-
-            // inline base64 dropped (no store), url kept (no store needed)
-            assertEquals(1, gc.getAttachments().size());
-            assertEquals("https://example.com/y.png", gc.getAttachments().get(0).getUrl());
-        }
-
-        @Test
-        void materialize_nullOrEmpty_noop() {
-            service.attachmentStore = mock(IAttachmentStore.class);
-            var gc = gc("gc-1");
-            service.materializeAttachments(gc, null);
-            service.materializeAttachments(gc, List.of());
-            assertNull(gc.getAttachments());
-        }
-
-        @Test
-        void grantAndInject_storedRef_grantsAndInjects() throws Exception {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            var gc = gc("gc-1");
-            gc.setAttachments(List.of(new Attachment("application/pdf", "doc.pdf", 10, "ref-1")));
-            Map<String, Context> context = new LinkedHashMap<>();
-
-            service.grantAndInjectAttachments(gc, "member-conv", context);
-
-            verify(store).grantAccess("ref-1", "member-conv");
-            assertTrue(context.containsKey("attachment_0"));
-            var value = (Map<?, ?>) context.get("attachment_0").getValue();
-            assertEquals("ref-1", value.get("storageRef"));
-            assertEquals("doc.pdf", value.get("fileName"));
-        }
-
-        @Test
-        void grantAndInject_url_injectsWithoutGrant() {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            var gc = gc("gc-1");
-            var url = new Attachment();
-            url.setMimeType("image/png");
-            url.setUrl("https://example.com/y.png");
-            gc.setAttachments(List.of(url));
-            Map<String, Context> context = new LinkedHashMap<>();
-
-            service.grantAndInjectAttachments(gc, "member-conv", context);
-
-            verifyNoInteractions(store);
-            var value = (Map<?, ?>) context.get("attachment_0").getValue();
-            assertEquals("https://example.com/y.png", value.get("url"));
-        }
-
-        @Test
-        void grantAndInject_grantFailure_skipsEntry() throws Exception {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            doThrow(new IAttachmentStore.AttachmentStoreException("nope")).when(store).grantAccess(any(), any());
-            var gc = gc("gc-1");
-            gc.setAttachments(List.of(new Attachment("application/pdf", "d.pdf", 1, "ref-1")));
-            Map<String, Context> context = new LinkedHashMap<>();
-
-            service.grantAndInjectAttachments(gc, "m", context);
-            assertFalse(context.containsKey("attachment_0"));
-        }
-
-        @Test
-        void grantAndInject_noAttachments_noop() {
-            service.attachmentStore = mock(IAttachmentStore.class);
-            Map<String, Context> context = new LinkedHashMap<>();
-            service.grantAndInjectAttachments(gc("gc-1"), "m", context);
-            assertTrue(context.isEmpty());
-        }
-
-        // rehydrateAttachmentsFromStore — recovers the transient attachments list
-        // after a HITL resume reloads the GroupConversation (attachments are
-        // @JsonIgnore transient, so a reloaded GC has none).
-
-        @Test
-        void rehydrate_nullAttachments_rebuildsFromStore() {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            when(store.listByConversation("gc-1")).thenReturn(
-                    List.of(new IAttachmentStore.Attachment("ref-1", "doc.pdf", "application/pdf", 10, "gc-1")));
-            var gc = gc("gc-1");
-
-            service.rehydrateAttachmentsFromStore(gc);
-
-            assertEquals(1, gc.getAttachments().size());
-            assertEquals("ref-1", gc.getAttachments().get(0).getStorageRef());
-            assertEquals("doc.pdf", gc.getAttachments().get(0).getFileName());
-            assertEquals("application/pdf", gc.getAttachments().get(0).getMimeType());
-            assertEquals(10L, gc.getAttachments().get(0).getSizeBytes());
-        }
-
-        @Test
-        void rehydrate_attachmentsAlreadyPresent_noop() {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            var gc = gc("gc-1");
-            gc.setAttachments(List.of(new Attachment("image/png", "keep.png", 5, "ref-keep")));
-
-            service.rehydrateAttachmentsFromStore(gc);
-
-            // fresh discussion already has its list — the store must not be consulted
-            verifyNoInteractions(store);
-            assertEquals(1, gc.getAttachments().size());
-            assertEquals("ref-keep", gc.getAttachments().get(0).getStorageRef());
-        }
-
-        @Test
-        void rehydrate_emptyStore_leavesAttachmentsNull() {
-            var store = mock(IAttachmentStore.class);
-            service.attachmentStore = store;
-            when(store.listByConversation("gc-1")).thenReturn(List.of());
-
-            var gc = gc("gc-1");
-            service.rehydrateAttachmentsFromStore(gc);
-
-            assertNull(gc.getAttachments());
-        }
-
-        @Test
-        void rehydrate_noStore_noop() {
-            service.attachmentStore = null;
-            var gc = gc("gc-1");
-            service.rehydrateAttachmentsFromStore(gc);
-            assertNull(gc.getAttachments());
-        }
-    }
 
     // =================================================================
     // discuss() tests
